+++ pandas_ml_utils/classification/summary.py.html
@@ -1,5 +1,5 @@
 <%
-    from pandas_ml_utils.utils import fig_to_png_base64
+    from pandas_ml_utils.utils.functions import fig_to_png_base64
 
     interpretation = {
         "FP Ratio": "FN/TP Ratio (should be < 0.5, ideally 0)",
